requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
google-generativeai==0.3.2
requests==2.31.0
